adds new gtfs static/realtime integration to notificaiton
