+++ app/src/main/java/ca/wheresthebus/utils/StaticDataLoadHelper.kt
@@ -1,10 +1,6 @@
 package ca.wheresthebus.utils
 
-import android.Manifest
 import android.content.Context
-import android.content.pm.PackageManager
-import android.widget.Toast
-import androidx.core.app.ActivityCompat
 import ca.wheresthebus.R
 import ca.wheresthebus.data.json.JsonRoute
 import ca.wheresthebus.data.json.JsonStop
@@ -131,20 +127,6 @@ object StaticDataLoadHelper {
             }
             val end = System.currentTimeMillis()
             println("Populating database with stop and route data took ${(end - start) / 1000} seconds")
-        }
-    }
-
-    fun checkLocationPermission(context: Context): Boolean {
-        if (ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_FINE_LOCATION
-        ) == PackageManager.PERMISSION_GRANTED &&
-                ActivityCompat.checkSelfPermission(
-                    context,
-                    Manifest.permission.ACCESS_COARSE_LOCATION
-                ) == PackageManager.PERMISSION_GRANTED) {
-            return true
-        } else {
-            Toast.makeText(context, "Location permission required", Toast.LENGTH_SHORT).show()
-            return false
         }
     }
 }
